backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
SQLAlchemy>=2.0.23
alembic>=1.12.1
mysqlclient>=2.2.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
aiosmtplib>=2.0.2
httpx>=0.25.1
Pillow>=10.1.0
click>=8.1.7 
Pillow==10.2.0